+++ src/reactive/effect-context.ts
@@ -77,7 +77,6 @@ export function getActiveEffect() {
  * @returns The result of the function
  */
 export function withEffect<T>(effect: ScopedCallback | undefined, fn: () => T): T {
-
 	if (getRoot(effect) === getRoot(getActiveEffect())) return fn()
 	stack.unshift(effect)
 	try {