+++ src/reactive/set.ts
@@ -1,10 +1,8 @@
+import { contentRef } from '../utils'
 import { touched, touched1 } from './change'
 import { makeReactiveEntriesIterator, makeReactiveIterator } from './non-reactive'
 import { reactive } from './proxy'
 import { dependant } from './tracking'
-import { bunch } from './utils'
-
-const native = Symbol('native')
 
 /**
  * Reactive wrapper around JavaScript's WeakSet class
@@ -19,7 +17,7 @@ export abstract class ReactiveWeakSet<T extends object> extends WeakSet<T> {
 		this.add(value)
 		if (!had) {
 			// touch the specific value and the collection view
-			touched1(bunch(this), { type: 'add', prop: value }, value)
+			touched1(contentRef(this), { type: 'add', prop: value }, value)
 			// no size/allProps for WeakSet
 		}
 		return this
@@ -28,12 +26,12 @@ export abstract class ReactiveWeakSet<T extends object> extends WeakSet<T> {
 	delete(value: T): boolean {
 		const had = this.has(value)
 		const res = this.delete(value)
-		if (had) touched1(bunch(this), { type: 'del', prop: value }, value)
+		if (had) touched1(contentRef(this), { type: 'del', prop: value }, value)
 		return res
 	}
 
 	has(value: T): boolean {
-		dependant(bunch(this), value)
+		dependant(contentRef(this), value)
 		return this.has(value)
 	}
 }
@@ -60,7 +58,7 @@ export abstract class ReactiveSet<T> extends Set<T> {
 		if (!had) {
 			const evolution = { type: 'add', prop: reactiveValue } as const
 			// touch for value-specific and aggregate dependencies
-			touched1(bunch(this), evolution, reactiveValue)
+			touched1(contentRef(this), evolution, reactiveValue)
 			touched1(this, evolution, 'size')
 		}
 		return this
@@ -72,7 +70,7 @@ export abstract class ReactiveSet<T> extends Set<T> {
 		if (hadEntries) {
 			const evolution = { type: 'bunch', method: 'clear' } as const
 			touched1(this, evolution, 'size')
-			touched(bunch(this), evolution)
+			touched(contentRef(this), evolution)
 		}
 	}
 
@@ -81,39 +79,39 @@ export abstract class ReactiveSet<T> extends Set<T> {
 		const res = this.delete(value)
 		if (had) {
 			const evolution = { type: 'del', prop: value } as const
-			touched1(bunch(this), evolution, value)
+			touched1(contentRef(this), evolution, value)
 			touched1(this, evolution, 'size')
 		}
 		return res
 	}
 
 	has(value: T): boolean {
-		dependant(bunch(this), value)
+		dependant(contentRef(this), value)
 		return this.has(value)
 	}
 
 	entries(): Generator<[T, T]> {
-		dependant(bunch(this))
+		dependant(contentRef(this))
 		return makeReactiveEntriesIterator(this.entries())
 	}
 
 	forEach(callbackfn: (value: T, value2: T, set: Set<T>) => void, thisArg?: any): void {
-		dependant(bunch(this))
+		dependant(contentRef(this))
 		this.forEach(callbackfn, thisArg)
 	}
 
 	keys(): Generator<T> {
-		dependant(bunch(this))
+		dependant(contentRef(this))
 		return makeReactiveIterator(this.keys())
 	}
 
 	values(): Generator<T> {
-		dependant(bunch(this))
+		dependant(contentRef(this))
 		return makeReactiveIterator(this.values())
 	}
 
 	[Symbol.iterator](): SetIterator<T> {
-		dependant(bunch(this))
+		dependant(contentRef(this))
 		const nativeIterator = this[Symbol.iterator]()
 		return {
 			next() {